rafctor copy script into eleventy.js
